feat(counties) create counties route
