+++ server/src/ClientToDBAPI/mainRoute.ts
@@ -10,6 +10,7 @@ import contactedPeopleRoute from './ContactedPeople/mainRoute';
 import intersectionsRoute from './IntersectionsRoute/mainRoute';
 import personalDetailsRoute from './PersonalDetailsRoute/mainRoute';
 import clinicalDetailsRoute from './ClinicalDetailsRoute/mainRoute';
+import countiesRoute from './CountiesRoute/mainRoute';
 
 const clientToDBRouter = Router();
 clientToDBRouter.use(authMiddleware);
@@ -22,5 +23,6 @@ clientToDBRouter.use('/clinicalDetails', clinicalDetailsRoute);
 clientToDBRouter.use('/addressDetails', addressRoute);
 clientToDBRouter.use('/contactedPeople', contactedPeopleRoute);
 clientToDBRouter.use('/users', usersRoute);
+clientToDBRouter.use('/counties', countiesRoute);
 
 export default clientToDBRouter;
