+++ server/src/DBService/Counties/Query.ts
@@ -1,0 +1,9 @@
+import {gql} from 'postgraphile';
+
+export const GET_COUNTY_DISPLAY_NAME_BY_USER = gql`
+query getCounty($id: Int!) {
+  countyById(id: $id) {
+    displayName
+  }
+}
+`;
