+++ server/src/DBService/Users/Query.ts
@@ -36,11 +36,3 @@ query AllGroupUsers($investigationGroup: Int!) {
   }
 }
 `;
-
-export const GET_COUNTY_BY_USER = gql`
-query getCounty($id: Int!) {
-  countyById(id: $id) {
-    displayName
-  }
-}
-`;
